create cifar example
